+++ src/main/java/io/github/karols/hocr4j/dom/HocrParser.java
@@ -22,7 +22,7 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.Queue;
 
-public class HocrParser {
+public final class HocrParser {
 
     private HocrParser(){}