+++ doc/CHANGELOG.md
@@ -1,13 +1,37 @@
 Changelog
 =========
 
+0.1.2
+-----
+
+* FIXED: `pom.xml` is now valid
+
+* FIXED: `Page.growBoundsForLeftAlignedColumn` now works
+
+* **Changed semantics of `Bounds.isEmpty` and `Bounds.intersects`**
+
+* Comparators in `OrderedBy` now return 0 less often
+
+* Added `@Immutable` annotation to model classes
+
+* Added `translate` method to most model classes
+
+* `PageRenderer` has configurable output file format now
+
+* Apart from that, added 3 methods to the `Bounds` class,
+1 method to the `Line` class,
+1 method to the `Page` class,
+3 methods to the `LineThat` class
+
+* Various style-related fixes
+
 0.1.1
 -----
 
 * Added 19 methods to the `Bounds` class,
 3 methods to the `Line` class,
-2 methods to the `Page` class,
 1 method to the `Paragraph` class,
+2 methods to the `Page` class,
 3 methods to the `TextUtils` class
 
 * Changed `toString` for most classes